Add configurable options menu via OptionsScreenMixin
Introduces OptionsScreenMixin to allow dynamic configuration of the options menu buttons based on a new OptionsConfig JSON file. Updates CustomMenuHandler to load and reference the options config, and registers the mixin in the mixins config. This enables users to control the visibility of various options menu buttons through a simple config file.